split out listener init
